+++ tools/eslint-plugin-sprout/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @qlik/eslint-plugin-sprout
 
+## 1.1.0
+
+### Minor Changes
+
+- 81db710: feat: add rule to forbid import of tokens
+
+### Patch Changes
+
+- 5328fb1: chore: update npm dependencies
+
+## 1.0.3
+
+### Patch Changes
+
+- 7c52fd0: chore: update npm dependencies
+- 9de9acf: chore: update deps
+
+## 1.0.2
+
+### Patch Changes
+
+- 0f95ab1: docs: update license to MIT
+
+## 1.0.1
+
+### Patch Changes
+
+- 906e787: chore: update npm dependencies
+
 ## 1.0.0
 
 ### Major Changes